add test case for integrity part
